greenhires.com-BE/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["tgiang352001 <t.giang352001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.1"
pydantic-settings = "^2.1.0"
fastapi = "^0.109.2"
redis = "^5.0.1"
sentry-sdk = "^1.39.1"
python-jose = "^3.3.0"
bcrypt = "^4.1.2"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
asyncpg = "^0.29.0"
pydantic = { extras = ["email"], version = "^2.5.3" }
python-multipart = "^0.0.6"
httpx = "^0.26.0"
oauthlib = "^3.2.2"
fastapi-mail = "^1.4.1"
boto3 = "^1.34.10"
pytest-playwright = "^0.4.3"
pypdf = "^3.17.4"
tenacity = "^8.2.3"
orjson = "^3.9.13"
pyotp = "^2.9.0"
gunicorn = "^21.2.0"
python-json-logger = "^2.0.7"
sse-starlette = "^2.1.0"
fastapi-limiter = "^0.1.6"
asgiref = "^3.8.1"
xdbsearchip = "^1.0.2"
apscheduler = "^3.10.4"
psycopg2-binary = "^2.9.9"
html2text = "^2024.2.26"
aiofiles = "^23.2.1"
openai = "^1.28.2"
pydantic-extra-types = "^2.7.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.30" }
pdftotext = "^2.2.2"
paypalrestsdk = "^1.13.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
